refactor: profile interface 수정
types/api -> types/profile
