+++ frontend/src/types/api.ts
@@ -12,19 +12,3 @@ export interface INoContent {
   isSuccess: boolean;
   message: string;
 }
-
-export interface IProfile {
-  id: number;
-  email: string;
-  imageUrl: string;
-  address: string;
-  nickName: string;
-  mobileNumber: string;
-}
-
-export interface IProfileEdit {
-  email: string;
-  address: string;
-  nickName: string;
-  mobileNumber: string;
-}
